[Improved] Gizmo (still bugged), [Fix] operator = with Vector2, [Fix] ScreenPosToRay + mouse position on sceneWindow

+++ GalaxyEngine/include/EditorUI/SceneWindow.h
@@ -32,8 +32,8 @@ namespace GALAXY
 			bool m_visible = true;
 			bool m_isHovered = false;
 
-			Vec2f m_imagePosition = 0;
-			Vec2f m_imageSize = 0;
+			Vec2i m_imagePosition = 0;
+			Vec2i m_imageSize = 0;
 
 			std::weak_ptr<Resource::Texture> m_settingsIcon;
 			std::weak_ptr<Resource::Texture> m_menuIcon;
